Refactored SIS Scraper
