+++ README.md
@@ -11,4 +11,10 @@ npm run dev
 
 This was done on my mac and hopefully you have your dependencies installed . . . which I didnt keep track of. Hopefully npm install does all the dependencies stuff for you 
 
+# Backend Setup
 
+```
+python3 -m venv .venv
+source .venv/bin/activate
+pip3 install requirements.txt
+```
