--- SeleniumScraper/main.py
@@ -1,11 +1,0 @@
-from Excel_Reader import Excel_Reader
-from SIS_Scraper import SIS_Scraper
-
-# term = "Fall 2023"
-# searchMe = "CSDS132"
-#
-# scraper = SIS_Scraper(term, searchMe)
-# scraper.search()
-q = Excel_Reader("test")
-
-
