+++ backend/main.py
@@ -1,0 +1,8 @@
+from excel_reader import ExcelReader
+from sis_scraper import SISScraper
+
+scraper = SISScraper()
+scraper.search("Fall 2023", "CSDS 132")
+# q = ExcelReader("test")
+
+
