+++ requirements.txt
@@ -1,0 +1,2 @@
+selenium==4.11.2
+pandas==2.1.1
